Get observations time series

Get available time series for observations.

Usage

get_observations_ts(
  sources = NULL,
  reference_time = NULL,
  elements = NULL,
  time_offsets = NULL,
  time_resolutions = NULL,
  time_series_ids = NULL,
  performance_categories = NULL,
  exposure_categories = NULL,
  qualities = NULL,
  levels = NULL,
  level_types = NULL,
  level_units = NULL,
  include_extra = NULL,
  fields = NULL,
  version = "v0",
  format = "jsonld",
  client = get_frost_client(),
  flatten = TRUE,
  return_response = FALSE
)

Arguments

sources

character: Source (station) ID, e.g. 'SN18700'.

reference_time

character: Time range to get observations for, in ISO-8601 format.

elements

character: Elements. See `get_elements()` for available values.

time_offsets

character: The time offsets to get observations for, specified as ISO-8601 periods.

time_resolutions

character: The time resolutions to get observations for, specified as ISO-8601 periods.

time_series_ids

integer: The internal time series IDs to get observations for. Defaults to 0.

performance_categories

character: The performance categories to get observations, e.g "A", "B" etc.

exposure_categories

integer: The exposure categories to get observations for, e.g. 1, 2 etc.

qualities

integer: The qualities to get observations for, e.g. 1, 2 etc.

levels

numeric: The sensor levels to get observations for, e.g. 0.1, 2, 10, 20.

level_types

character: Sensor level types

level_units

character: Sensor level units

include_extra

integer: If 1 extra data is returned.

fields

character: Fields to return in the response.

version

character: API version.

format

character: Response format.

client

list: List with client id and secret. Defaults to the `MET_FROST_ID` and `MET_FROST_SECRET` environment variables.

flatten

logical: If TRUE the response is transformed to a table.

return_response

logical: If TRUE a list of class `frost_api` is returned, including the raw `httr2_response`.

Details

Use the default query parameters to retrieve all time series.

Examples

if (FALSE) {
df <- get_observations_ts("SN23670")
}